# public/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Source Library API",
    "description": "Search and cite rare historical texts (alchemy, Hermeticism, Renaissance philosophy) from Source Library. All translations include DOI-backed academic citations.",
    "version": "1.1.0",
    "contact": {
      "name": "Source Library",
      "url": "https://sourcelibrary.org",
      "email": "team@sourcelibrary.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://sourcelibrary.org/api",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/dts/collection": {
      "get": {
        "summary": "DTS collections",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Collection or resource id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Dataset API key (sl_data_...). Free at https://sourcelibrary.org/developers."
      }
    }
  }
}